Better numerical approximation of absolute humidity
The previous implementation of the calculation of absolute humidity
suffered from an approximation error when the relative humidity was
close to 100%RH. It approximated a division by 100 by multiplying by 82
and then right shifting by 13. The new implementation has a lower
approximation error as it does not use the previous division
approximation.