Make definition of sgp40_sources consistent
The sgp40_sources variable was defined differently in the default
config of the sgp40 and the sgp40_voc_index part of this repo.
This is normally not a problem but in the test's Makefile both
configs are included and now the order of the includes is relevant.
If the sgp40 config was included last nothing happened but the other
way around the compilation of the sgp40 was missing the sensirion
common and the sgp common files and failed. This was not the case
in the master branch but an internal one, where the includes happened
to added in the wrong order.